+++ src/main/java/me/drex/worldmanager/gui/list/WorldList.java
@@ -1,7 +1,6 @@
 package me.drex.worldmanager.gui.list;
 
 import eu.pb4.sgui.api.elements.GuiElementBuilder;
-import me.drex.worldmanager.gui.create.SelectChunkGenerator;
 import me.drex.worldmanager.gui.util.PagedGui;
 import me.drex.worldmanager.save.WorldConfig;
 import me.drex.worldmanager.save.WorldManagerSavedData;
@@ -9,12 +8,8 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.inventory.MenuType;
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.level.chunk.ChunkGenerator;
-import net.minecraft.world.level.levelgen.DebugLevelSource;
-import net.minecraft.world.level.levelgen.NoiseBasedChunkGenerator;
-import xyz.nucleoid.fantasy.util.VoidChunkGenerator;
 
 import java.util.List;
 import java.util.Map;
@@ -35,18 +30,16 @@ public class WorldList extends PagedGui<Map.Entry<ResourceLocation, WorldConfig>
 
     @Override
     protected GuiElementBuilder toGuiElement(Map.Entry<ResourceLocation, WorldConfig> entry) {
-        return new GuiElementBuilder(toIcon(entry.getValue().generator))
-            .setName(Component.literal(entry.getKey().toString()));
+        WorldConfig config = entry.getValue();
+        ItemStack icon = config.data.icon;
+        GuiElementBuilder builder;
+        if (icon.isEmpty()) {
+            builder = new GuiElementBuilder(Items.PLAYER_HEAD)
+                .setSkullOwner("eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvYmFkYzA0OGE3Y2U3OGY3ZGFkNzJhMDdkYTI3ZDg1YzA5MTY4ODFlNTUyMmVlZWQxZTNkYWYyMTdhMzhjMWEifX19");
+        } else {
+            builder = new GuiElementBuilder(icon);
+        }
+        return builder.setName(Component.literal(entry.getKey().toString()));
     }
 
-    public static Item toIcon(ChunkGenerator chunkGenerator) {
-        if (chunkGenerator instanceof NoiseBasedChunkGenerator noiseBasedChunkGenerator) {
-            return noiseBasedChunkGenerator.generatorSettings().value().defaultBlock().getBlock().asItem();
-        } else if (chunkGenerator instanceof VoidChunkGenerator) {
-            return Items.STRUCTURE_VOID;
-        } else if (chunkGenerator instanceof DebugLevelSource) {
-            return Items.COMMAND_BLOCK;
-        }
-        return Items.STONE;
-    }
 }